secure-messaging/requirements.txt
cryptography
pycryptodome
pytest